added parallel routing to admin ui
