+++ src/app/admin/@comments/error.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+export default function CommentsError() {
+    return (
+        <h1>Error Comments!</h1>
+    )
+}
