+++ src/app/admin/@comments/loading.tsx
@@ -1,0 +1,5 @@
+export default function LoadingComments() {
+    return (
+        <h1>Loading Comments...</h1>
+    )
+}
