+++ src/app/admin/@feedback/error.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+export default function FeedbackError() {
+    return (
+        <h1>Error Feedback!</h1>
+    )
+}
