+++ src/app/admin/@feedback/loading.tsx
@@ -1,0 +1,5 @@
+export default function LoadingFeedback() {
+    return (
+        <h1>Loading Feedback...</h1>
+    )
+}
