+++ src/app/admin/@newblog/error.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+export default function NewBlogError() {
+    return (
+        <h1>Error with creating new blog post!</h1>
+    )
+}
