+++ src/app/admin/@newblog/loading.tsx
@@ -1,0 +1,5 @@
+export default function LoadingNewBlog() {
+    return (
+        <h1>Loading New Blog Form...</h1>
+    )
+}
